Added Account Creation html Page + jsp connection to mysql db
I added the form that allows users to create their account (connects to login page). I have also added an additional css file, along with the .jsp file to connect the new page with the mysql database in order to insert into the 'users' table. Note that this requires the database schema on your workbench to be named sjsu-bq.